Weekly user

See merge request 1dv508/student/the-caves-of-steel/cookbook!70

+++ app/src/main/java/cookbook/backend/be_objects/CommentObject.java
@@ -4,11 +4,11 @@ package cookbook.backend.be_objects;
 public class CommentObject {
   private int commentId;
   private String recipeId;
-  private int userId;
+  private Long userId;
   private String text;
   private String timestamp;
 
-  public CommentObject(int commentId, String recipeId, int userId, String text, String timestamp) {
+  public CommentObject(int commentId, String recipeId, Long userId, String text, String timestamp) {
       this.commentId = commentId;
       this.recipeId = recipeId;
       this.userId = userId;
@@ -28,10 +28,10 @@ public int getCommentId() {
   public void setRecipeId(String recipeId) {
       this.recipeId = recipeId;
   }
-  public int getUserId() {
+  public Long getUserId() {
       return userId;
   }
-  public void setUserId(int userId) {
+  public void setUserId(Long userId) {
       this.userId = userId;
   }
   public String getText() {
